+++ src/js/core/navbar.js
@@ -12,6 +12,7 @@ export default {
     },
 
     data: {
+        delayShow: 200,
         clsDrop: 'uk-navbar-dropdown',
         selNavItem:
             '.uk-navbar-nav > li > a,a.uk-navbar-item,button.uk-navbar-item,.uk-navbar-item a,.uk-navbar-item button,.uk-navbar-toggle', // Simplify with :where() selector once browser target is Safari 14+