+++ src/js/core/toggle.js
@@ -1,10 +1,10 @@
 import {
-    attr,
     hasAttr,
     hasClass,
     includes,
     isBoolean,
     isFocusable,
+    isSameSiteAnchor,
     isTag,
     isTouch,
     matches,
@@ -54,10 +54,10 @@ export default {
     connected() {
         if (!includes(this.mode, 'media')) {
             if (!isFocusable(this.$el)) {
-                attr(this.$el, 'tabindex', '0');
+                this.$el.tabIndex = 0;
             }
             if (!this.cls && isTag(this.$el, 'a')) {
-                attr(this.$el, 'role', 'button');
+                this.$el.role = 'button';
             }
         }
     },
@@ -154,18 +154,23 @@ export default {
             filter: ({ mode }) => ['click', 'hover'].some((m) => includes(mode, m)),
 
             handler(e) {
-                let link;
-                if (
-                    this._preventClick ||
-                    e.target.closest('a[href="#"], a[href=""]') ||
-                    ((link = e.target.closest('a[href]')) &&
-                        (!this.isToggled(this.target) ||
-                            (link.hash && matches(this.target, link.hash))))
-                ) {
+                if (e.defaultPrevented) {
+                    return;
+                }
+
+                const link = e.target.closest('a[href]');
+                const isButtonLike =
+                    isSameSiteAnchor(link) && (!link.hash || matches(this.target, link.hash));
+
+                if (this._preventClick || isButtonLike || (link && !this.isToggled(this.target))) {
                     e.preventDefault();
                 }
 
-                if (!this._preventClick && includes(this.mode, 'click')) {
+                if (
+                    !this._preventClick &&
+                    includes(this.mode, 'click') &&
+                    (!link || isButtonLike || e.defaultPrevented)
+                ) {
                     this.toggle();
                 }
             },
@@ -193,7 +198,7 @@ export default {
             }
 
             if (hasAttr(this.$el, 'aria-expanded')) {
-                attr(this.$el, 'aria-expanded', !this.isToggled(this.target));
+                this.$el.ariaExpanded = !this.isToggled(this.target);
             }
 
             if (!this.queued) {