+++ src/js/mixin/event.js
@@ -1,0 +1,5 @@
+export function maybeDefaultPreventClick(e) {
+    if (e.target.closest('a[href="#"],a[href=""]')) {
+        e.preventDefault();
+    }
+}